CSV memmap dataloader (#42)
* Debugging

* Testing data loading cache change

* Lazy loading implemented

* Modified version of mem-mapped dataloader...

The original mem-loader from Micha has following changes:
 - Remove dask array, which as reduced the loading time to fraction of original
 - Multiprocessing to build mem-mapped file

* Remove the hardcoding for column selection

This is a temp fix. Ideally, dataset should return the complete line.
However, using this WAR to limit the changes to just this component.

* Fix distributed index build issue...

Previously in a smaller setup, I had noticed that only one node was building the indexes
and others were just waiting while the indexes were created. Micha, mentioned that it is
not a valid assumption. Hence, the rank contraint and barrier.

Other changes:
    - Build index function is moved to an util script

* Minor bugfixes for syntax error/variable scope issues

Also refactored utils. to move a local function outside the parent
function.

* Rename dataset class

* Rename dataset class

* Fix Syntax errors

* 1. Debugging.

Signed-off-by: Micha Livne <mlivne@nvidia.com>

* 1. Debugging.

Signed-off-by: Micha Livne <mlivne@nvidia.com>

* 1. Debugging.

Signed-off-by: Micha Livne <mlivne@nvidia.com>

* 1. Added logging instead of print.

Signed-off-by: Micha Livne <mlivne@nvidia.com>

Co-authored-by: Michelle Gill <mgill@nvidia.com>
Co-authored-by: Rajesh K Ilango <rilango@nvidia.com>
Co-authored-by: Micha Livne <mlivne@nvidia.com>